17 Desember 2024

+++ A4/GameEngine.hpp
@@ -27,7 +27,7 @@ protected:
 	void                   update();
 	void                   sUserInput();
 
-	float currentFrame = 0.0f;
+	int currentFrame = 0;
 
 public:
 	sf::Clock deltaClock;
